requirements.txt
flask==3.1.1
anthropic==0.52.0
pdf2image==1.17.0
Pillow==11.1.0
pytesseract==0.3.13
openpyxl==3.1.5
python-dotenv==1.1.0
werkzeug==3.1.3
